+++ lib/instream/connection/query_runner.ex
@@ -3,6 +3,11 @@ defmodule Instream.Connection.QueryRunner do
   Executes a query for a connection.
   """
 
+  alias Instream.Log.Metadata
+  alias Instream.Log.PingEntry
+  alias Instream.Log.QueryEntry
+  alias Instream.Log.StatusEntry
+  alias Instream.Log.WriteEntry
   alias Instream.Query
   alias Instream.Query.Headers
   alias Instream.Query.URL
@@ -12,13 +17,33 @@ defmodule Instream.Connection.QueryRunner do
   Executes `:ping` queries.
   """
   @spec ping(Query.t, Keyword.t, Keyword.t) :: :pong | :error
-  def ping(%Query{} = query, _opts, conn) do
+  def ping(%Query{} = query, opts, conn) do
     headers = conn |> Headers.assemble()
 
-    conn
-    |> URL.ping(query.opts[:host])
-    |> :hackney.head(headers, "", Keyword.get(conn, :http_opts, []))
-    |> Response.parse_ping()
+    { query_time, response } = :timer.tc fn ->
+      conn
+      |> URL.ping(query.opts[:host])
+      |> :hackney.head(headers, "", Keyword.get(conn, :http_opts, []))
+    end
+
+    result = response |> Response.parse_ping()
+    status = case response do
+      { :ok, status, _ } -> status
+      _                  -> 0
+    end
+
+    if false != opts[:log] do
+      conn[:module].__log__(%PingEntry{
+        host:     query.opts[:host] || hd(conn[:hosts]),
+        result:   result,
+        metadata: %Metadata{
+          query_time:      query_time,
+          response_status: status
+        }
+      })
+    end
+
+    result
   end
 
   @doc """
@@ -28,7 +53,7 @@ defmodule Instream.Connection.QueryRunner do
   def read(%Query{} = query, opts, conn) do
     headers = conn |> Headers.assemble()
     url     =
-         conn
+      conn
       |> URL.query()
       |> URL.append_database(opts[:database])
       |> URL.append_epoch(query.opts[:precision])
@@ -36,23 +61,59 @@ defmodule Instream.Connection.QueryRunner do
 
     http_opts = Keyword.get(conn, :http_opts, [])
 
-    { :ok, status, headers, client } = :hackney.get(url, headers, "", http_opts)
-    { :ok, response }                = :hackney.body(client)
+    { query_time, { :ok, status, headers, client }} = :timer.tc fn ->
+      :hackney.get(url, headers, "", http_opts)
+    end
 
-    { status, headers, response } |> Response.maybe_parse(opts)
+    { :ok, response } = :hackney.body(client)
+    result            =
+      { status, headers, response }
+      |> Response.maybe_parse(opts)
+
+    if false != opts[:log] do
+      conn[:module].__log__(%QueryEntry{
+        query:    query.payload,
+        metadata: %Metadata{
+          query_time:      query_time,
+          response_status: status
+        }
+      })
+    end
+
+    result
   end
 
   @doc """
   Execute `:status` queries.
   """
   @spec status(Query.t, Keyword.t, Keyword.t) :: :ok | :error
-  def status(%Query{} = query, _opts, conn) do
+  def status(%Query{} = query, opts, conn) do
     headers = conn |> Headers.assemble()
 
-    conn
-    |> URL.status(query.opts[:host])
-    |> :hackney.head(headers, "", Keyword.get(conn, :http_opts, []))
-    |> Response.parse_status()
+    { query_time, response } = :timer.tc fn ->
+      conn
+      |> URL.status(query.opts[:host])
+      |> :hackney.head(headers, "", Keyword.get(conn, :http_opts, []))
+    end
+
+    result = response |> Response.parse_status()
+    status = case response do
+      { :ok, status, _ } -> status
+      _                  -> 0
+    end
+
+    if false != opts[:log] do
+      conn[:module].__log__(%StatusEntry{
+        host:     query.opts[:host] || hd(conn[:hosts]),
+        result:   result,
+        metadata: %Metadata{
+          query_time:      query_time,
+          response_status: status
+        }
+      })
+    end
+
+    result
   end
 
   @doc """
@@ -60,8 +121,22 @@ defmodule Instream.Connection.QueryRunner do
   """
   @spec write(Query.t, Keyword.t, Keyword.t) :: any
   def write(%Query{} = query, opts, conn) do
-    query
-    |> conn[:writer].write(opts, conn)
-    |> Response.maybe_parse(opts)
+    { query_time, result } = :timer.tc fn ->
+      query
+      |> conn[:writer].write(opts, conn)
+      |> Response.maybe_parse(opts)
+    end
+
+    if false != opts[:log] do
+      conn[:module].__log__(%WriteEntry{
+        points:   length(query.payload[:points]),
+        metadata: %Metadata{
+          query_time:      query_time,
+          response_status: 0
+        }
+      })
+    end
+
+    result
   end
 end