+++ test/helpers/nil_logger.exs
@@ -1,0 +1,3 @@
+defmodule Instream.TestHelpers.NilLogger do
+  def log(entry), do: entry
+end