+++ test/test_helper.exs
@@ -1,6 +1,9 @@
+Code.require_file("helpers/nil_logger.exs", __DIR__)
+
 Code.require_file("helpers/connection.exs", __DIR__)
 Code.require_file("helpers/connection_with_opts.exs", __DIR__)
 Code.require_file("helpers/json_connection.exs", __DIR__)
+Code.require_file("helpers/log_connection.exs", __DIR__)
 Code.require_file("helpers/udp_connection.exs", __DIR__)
 
 Code.require_file("helpers/anon_connection.exs", __DIR__)
@@ -19,6 +22,7 @@ alias Instream.TestHelpers
   TestHelpers.Connection.child_spec,
   TestHelpers.ConnectionWithOpts.child_spec,
   TestHelpers.JSONConnection.child_spec,
+  TestHelpers.LogConnection.child_spec,
   TestHelpers.UDPConnection.child_spec,
 
   TestHelpers.AnonConnection.child_spec,